requirements.txt
pandas==1.0.3
psycopg2-binary==2.8.4
requests==2.23.0
